ProblemSet1B_HouseHunting_WithaRaise
Write a program to calculate how many months it will take you to save up
enough money for a down payment. ProblemSet1A modified to include the
following:
1.Have the user input a semi-annual salary raise semi_annual_raise (as a
decimal percentage)
2.After the 6 th month, increase your salary by that percentage. Do the
same after the 12th th month, the 18 month, and so on.
The program print results in the format shown in the test cases below:
Test Case 1
>>>
Enter your annual salary: 120000
Enter the percent of your salary to save, as a decimal: . 05
Enter the cost of your dream home: 500000
Enter the semiannual raise, as a decimal: .03
Number of months: 142
>>>
Test Case 2
>>>
Enter your annual salary: 80000
Enter the percent of your salary to save, as a decimal: . 1
Enter the cost of your dream home: 800000
Enter the semiannual raise, as a decimal: .03
Number of months: 159
>>>
Test Case 3
>>>
Enter your annual salary: 75000
Enter the percent of your salary to save, as a decimal: . 05
Enter the cost of your dream home: 1500000
Enter the semiannual raise, as a decimal: .05
Number of months: 261
>>>