ProblemSet1C
How much should you save each month to to be able to afford the down
payment of a house in three years? In this problem, you are going to
write a program to answer that question. To simplify things, assume:
1.Your semiannual raise is .07 (7%)
2.Your investments have an annual return of 0.04 (4%)
3.The down payment is 0.25 (25%) of the cost of the house
4.The cost of the house that you are saving for is $1M.
Program prints:
Test Case 1
>>>
Enter the starting salary: 150000
Best savings rate: 0.4411
Steps in bisection search: 12
>>>
Test Case 2
>>>
Enter the starting salary: 300000
Best savings rate: 0.2206
Steps in bisection search: 9
>>>
Test Case 3
>>>
Enter the starting salary: 10000
It is not possible to pay the down payment in three years.
>>>